requirements.txt
python-dotenv
boto3
flask
flask_restful
google-cloud-vision